+++ agent/internal/client/methods.go
